+++ Makefile
@@ -1,6 +1,6 @@
 check:
-	flake8 jmeslog.py
-	python -m mypy --disallow-untyped-defs --strict-optional --warn-no-return ./jmeslog.py
+	flake8 jmeslog
+	python -m mypy --disallow-untyped-defs --strict-optional --warn-no-return jmeslog
 
 coverage:
 	py.test --cov jmeslog --cov-report term-missing ./tests
