+++ tests/test_jmeslog.py
@@ -6,6 +6,9 @@ from unittest import mock
 from typing import Optional
 
 import jmeslog
+import jmeslog.cli
+import jmeslog.core
+import jmeslog.errors
 from jmeslog import model
 
 
@@ -45,14 +48,14 @@ def test_can_validate_allowed_values():
     schema = model.EntrySchema(type=['feature', 'bugfix'])
     entry = model.JMESLogEntry(type='feature', category='foo',
                                  description='My feature')
-    assert jmeslog.validate_change_entry(entry=entry, schema=schema) is None
+    assert jmeslog.core.validate_change_entry(entry=entry, schema=schema) is None
     err_msg = (
         'The "type" value must be one of: feature, bugfix, '
         'received: "notafeature"'
     )
-    with pytest.raises(jmeslog.ValidationError,
+    with pytest.raises(jmeslog.errors.ValidationError,
                        match=err_msg):
-        jmeslog.validate_change_entry(
+        jmeslog.core.validate_change_entry(
             schema=schema,
             entry=model.JMESLogEntry(
                 type='notafeature',
@@ -66,9 +69,9 @@ def test_no_values_can_be_empty():
     entry = model.JMESLogEntry(type='feature', category='foo',
                                  description='')
     err_msg = 'The "description" value cannot be empty.'
-    with pytest.raises(jmeslog.ValidationError,
+    with pytest.raises(jmeslog.errors.ValidationError,
                        match=err_msg):
-        jmeslog.validate_change_entry(
+        jmeslog.core.validate_change_entry(
             schema=schema,
             entry=entry,
         )
@@ -80,8 +83,8 @@ def test_no_prompt_if_entry_complete():
         category='foo',
         description='My Feature'
     )
-    retriever = mock.Mock(spec=jmeslog.EditorRetriever)
-    gen = jmeslog.EntryGenerator(completed, retriever)
+    retriever = mock.Mock(spec=jmeslog.core.EditorRetriever)
+    gen = jmeslog.core.EntryGenerator(completed, retriever)
     gen.complete_entry()
     assert not retriever.prompt_entry_values.called
 
@@ -93,8 +96,8 @@ def test_prompt_in_editor_if_incomplete():
         # Missing a description.
         description='',
     )
-    retriever = mock.Mock(spec=jmeslog.EditorRetriever)
-    gen = jmeslog.EntryGenerator(incomplete, retriever)
+    retriever = mock.Mock(spec=jmeslog.core.EditorRetriever)
+    gen = jmeslog.core.EntryGenerator(incomplete, retriever)
     gen.complete_entry()
     retriever.prompt_entry_values.assert_called_with(incomplete)
 
@@ -105,8 +108,8 @@ def test_can_return_generated_entry():
         category='foo',
         description='My Feature'
     )
-    retriever = mock.Mock(spec=jmeslog.EditorRetriever)
-    gen = jmeslog.EntryGenerator(completed, retriever)
+    retriever = mock.Mock(spec=jmeslog.core.EditorRetriever)
+    gen = jmeslog.core.EntryGenerator(completed, retriever)
     assert gen.change_entry.to_json() == (
         '{\n'
         '  "type": "feature",\n'
@@ -124,13 +127,13 @@ def test_can_record_entry(tmpdir):
     )
     change_dir = tmpdir.join('.changes')
     change_dir.mkdir()
-    recorder = jmeslog.EntryRecorder(
-        entry_gen=jmeslog.EntryGenerator(
+    recorder = jmeslog.core.EntryRecorder(
+        entry_gen=jmeslog.core.EntryGenerator(
             entry=entry,
-            retriever=mock.Mock(spec=jmeslog.EditorRetriever),
+            retriever=mock.Mock(spec=jmeslog.core.EditorRetriever),
         ),
         schema=model.EntrySchema(),
-        file_writer=jmeslog.EntryFileWriter(),
+        file_writer=jmeslog.core.EntryFileWriter(),
         output_dir=str(change_dir),
     )
     recorder.write_change_file_entry()
@@ -139,7 +142,7 @@ def test_can_record_entry(tmpdir):
 
 
 def _assert_entry_file_parses_to(contents, expected):
-    result = jmeslog.EntryFileParser().parse_contents(contents)
+    result = jmeslog.core.EntryFileParser().parse_contents(contents)
     assert result == expected
 
 
@@ -225,7 +228,7 @@ def test_collection_to_dict():
 
 def write_change(change_type, change_dir):
     entry = new_change(change_type)
-    jmeslog.create_entry_recorder(entry, change_dir).write_change_file_entry()
+    jmeslog.core.create_entry_recorder(entry, change_dir).write_change_file_entry()
     return entry
 
 
@@ -236,7 +239,7 @@ def test_can_load_next_changes_dir_into_entries(tmpdir):
     write_change('bugfix', str(change_dir))
     write_change('enhancement', str(change_dir))
 
-    entries = jmeslog.load_next_changes(str(change_dir))
+    entries = jmeslog.core.load_next_changes(str(change_dir))
     assert entries == model.JMESLogEntryCollection(
         changes=[new_change('feature'),
                  new_change('bugfix'),
@@ -259,7 +262,7 @@ def test_can_load_next_changes_dir_into_entries(tmpdir):
     ]
 )
 def test_determine_next_version(last_version, bump_type, new_version):
-    assert jmeslog.determine_next_version(
+    assert jmeslog.core.determine_next_version(
         last_version, bump_type) == new_version
 
 
@@ -274,7 +277,7 @@ def test_can_find_last_released_version(tmpdir):
     change_dir.join('1.2.30.json').write('{}')
     change_dir.join('1.10.0.json').write('{}')
     change_dir.join('1.20.0.json').write('{}')
-    assert jmeslog.find_last_released_version(str(change_dir)) == '1.20.0'
+    assert jmeslog.core.find_last_released_version(str(change_dir)) == '1.20.0'
 
 
 def test_can_consolidate_next_release(tmpdir):
@@ -293,7 +296,7 @@ def test_can_consolidate_next_release(tmpdir):
     changes.append(
         write_change('enhancement', next_release_dir)
     )
-    jmeslog.consolidate_next_release(
+    jmeslog.core.consolidate_next_release(
         next_version='1.1.0', change_dir=str(change_dir),
         changes=model.JMESLogEntryCollection(changes=changes))
     assert os.path.isfile(str(change_dir.join('1.1.0.json')))
@@ -342,13 +345,13 @@ def test_can_create_collection_from_old_format():
 def test_can_set_explicit_version(tmpdir):
     change_dir = os.path.join(str(tmpdir), '.changes')
     args = CommandArgs(change_dir=change_dir, release_version='1.2.3')
-    jmeslog.cmd_init(args)
+    jmeslog.cli.cmd_init(args)
     new_change_args = NewChangeArgs(
         type='enhancement',
         category='Foo',
         description='Changed foo',
         change_dir=change_dir,
     )
-    jmeslog.cmd_new_change(new_change_args)
-    jmeslog.cmd_new_release(args)
+    jmeslog.cli.cmd_new_change(new_change_args)
+    jmeslog.cli.cmd_new_release(args)
     assert os.listdir(change_dir)[0] == '1.2.3.json'
